headless: rename the nogui target and align target/source names
