+++ headless/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_executable(headless headless.cpp emulators.cpp settings.cpp)
+
+target_include_directories(headless PRIVATE ${CMAKE_SOURCE_DIR})
+target_compile_definitions(headless PRIVATE MIA_LIBRARY=1)
+target_link_libraries(headless PRIVATE ares::ares ares::mia ares::nall)
+
+if(ARES_ENABLE_CHD)
+  target_link_libraries(headless PRIVATE chdr-static)
+endif()
+
+set_target_properties(headless PROPERTIES OUTPUT_NAME ares-headless)
+
+ares_configure_executable(headless)
